refactor: extract CLI commands into separate modules
- Create cli_endpoints/ directory with modular command files
- Move evaluate commands to cli_endpoints/evaluate.py
- Move forecast commands to cli_endpoints/forecast.py
- Move utility commands to cli_endpoints/utils.py
- Extract shared helpers to cli_endpoints/_common.py
- Reduce cli.py to minimal entry point
- Update test imports to use new module locations

Commands are now top-level functions that can be imported directly,
making them easier to test and reuse programmatically.
