+++ chap_core/cli.py
@@ -1,670 +1,20 @@
 """Console script for chap_core."""
 
 import logging
-import dataclasses
-import json
-from pathlib import Path
-from typing import Literal, Optional, Any
 
-import numpy as np
-import pandas as pd
-import yaml
 from cyclopts import App
 
-from chap_core.api_types import BackTestParams, RunConfig
-from chap_core.assessment.dataset_splitting import train_test_generator
-from chap_core.assessment.evaluation import Evaluation
-from chap_core.assessment.forecast import multi_forecast as do_multi_forecast
-from chap_core.assessment.prediction_evaluator import evaluate_model
-from chap_core.database.model_templates_and_config_tables import ModelConfiguration
-from chap_core.datatypes import FullData
-from chap_core.exceptions import NoPredictionsError
-from chap_core.hpo.searcher import RandomSearcher
-from chap_core.models.model_template import ModelTemplate
-from chap_core.models.utils import (
-    get_model_from_directory_or_github_url,
-    get_model_template_from_directory_or_github_url,
-)
-from chap_core.geometry import Polygons
-from chap_core.log_config import initialize_logging
-from chap_core.plotting.dataset_plot import StandardizedFeaturePlot
-from chap_core.plotting.prediction_plot import plot_forecast_from_summaries
-from chap_core.plotting.season_plot import SeasonCorrelationBarPlot
-from chap_core.predictor import ModelType
-from chap_core.spatio_temporal_data.multi_country_dataset import (
-    MultiCountryDataSet,
-)
-from chap_core.spatio_temporal_data.temporal_dataclass import DataSet
-from chap_core import api, get_temp_dir
-from chap_core.file_io.example_data_set import datasets, DataSetType
-from chap_core.time_period.date_util_wrapper import delta_month
-
-from chap_core.hpo.hpoModel import HpoModel, Direction
-from chap_core.hpo.objective import Objective
-from chap_core.hpo.base import load_search_space_from_config
-
+from chap_core.cli_endpoints import evaluate, forecast, utils
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 app = App()
 
-
-def append_to_csv(file_object, data_frame: pd.DataFrame):
-    data_frame.to_csv(file_object, mode="a", header=False)
-
-
-@app.command()
-def evaluate_hpo(
-    model_name: ModelType | str,
-    dataset_name: Optional[DataSetType] = None,
-    dataset_country: Optional[str] = None,
-    dataset_csv: Optional[Path] = None,
-    polygons_json: Optional[Path] = None,
-    polygons_id_field: Optional[str] = "id",
-    prediction_length: int = 3,
-    n_splits: int = 7,
-    report_filename: Optional[str] = str(get_temp_dir() / "report.pdf"),
-    ignore_environment: bool = False,
-    debug: bool = False,
-    log_file: Optional[str] = None,
-    run_directory_type: Optional[Literal["latest", "timestamp", "use_existing"]] = "timestamp",
-    model_configuration_yaml: Optional[str] = None,
-    metric: Optional[str] = "MSE",
-    direction: Direction = "minimize",
-    evaluate_hpo: Optional[bool] = True,
-):
-    """
-    Same as evaluate, but has three added arguments and a if check on argument evaluate_hpo.
-    """
-    initialize_logging(debug, log_file)
-    if dataset_name is None:
-        assert dataset_csv is not None, "Must specify a dataset name or a dataset csv file"
-        logging.info(f"Loading dataset from {dataset_csv}")
-        dataset = DataSet.from_csv(dataset_csv, FullData)
-        if polygons_json is not None:
-            logging.info(f"Loading polygons from {polygons_json}")
-            polygons = Polygons.from_file(polygons_json, id_property=polygons_id_field)
-            polygons.filter_locations(dataset.locations())
-            dataset.set_polygons(polygons.data)
-    else:
-        logger.info(f"Evaluating model {model_name} on dataset {dataset_name}")
-
-        dataset = datasets[dataset_name]
-        dataset = dataset.load()
-
-        if isinstance(dataset, MultiCountryDataSet):
-            assert dataset_country is not None, "Must specify a country for multi country datasets"
-            assert dataset_country in dataset.countries, (
-                f"Country {dataset_country} not found in dataset. Countries: {dataset.countries}"
-            )
-            dataset = dataset[dataset_country]
-
-    if "," in model_name:
-        # model_name is not only one model, but contains a list of models
-        model_list = model_name.split(",")
-        model_configuration_yaml_list = [None for _ in model_list]
-        if model_configuration_yaml is not None:
-            model_configuration_yaml_list = model_configuration_yaml.split(",")
-            assert len(model_list) == len(model_configuration_yaml_list), (
-                "Number of model configurations does not match number of models"
-            )
-    else:
-        model_list = [model_name]
-        model_configuration_yaml_list = [model_configuration_yaml]
-
-    logging.info(f"Model configuration: {model_configuration_yaml_list}")
-
-    results_dict = {}
-    for name, configuration in zip(model_list, model_configuration_yaml_list):
-        template = ModelTemplate.from_directory_or_github_url(
-            name,
-            base_working_dir=Path("./runs/"),
-            ignore_env=ignore_environment,
-            run_dir_type=run_directory_type,
-        )
-        logging.info(f"Model template loaded: {template}")
-        if not evaluate_hpo:
-            if configuration is not None:
-                logger.info(f"Loading model configuration from yaml file {configuration}")
-                configuration = ModelConfiguration.model_validate(
-                    yaml.safe_load(open(configuration))
-                )  # template.get_model_configuration_from_yaml(Path(configuration))
-                logger.info(f"Loaded model configuration from yaml file: {configuration}")
-
-            model = template.get_model(configuration)
-            model = model()
-        else:
-            if configuration is not None:
-                logger.info(f"Loading model configuration from yaml file {configuration}")
-                with open(configuration, "r", encoding="utf-8") as f:
-                    configs = yaml.safe_load(f)
-                if not isinstance(configs, dict) or not configs:
-                    raise ValueError("YAML must define a non-empty mapping of parameters")
-                logger.info(f"Loaded model base configurations from yaml file: {configs}")
-            else:  # uses hpo_search_space defined in model's MLproject
-                configs = template.model_template_config.hpo_search_space
-
-            configs = load_search_space_from_config(configs)
-            objective = Objective(template, metric, prediction_length, n_splits)
-            model = HpoModel(RandomSearcher(2), objective, direction, configs)
-        try:
-            results = evaluate_model(
-                estimator=model,
-                data=dataset,
-                prediction_length=prediction_length,
-                n_test_sets=n_splits,
-                report_filename=report_filename,
-            )
-        except NoPredictionsError as e:
-            logger.error(f"No predictions were made: {e}")
-            return
-        print(f"Results: {results}")
-        results_dict[name] = results
-
-    # need to iterate through the dict, like key and value or something and then extract the relevant metrics
-    # to a pandas dataframe, and save it as a csv file.
-    # it seems like results contain two dictionairies, one for aggregate metrics and one with seperate ones for each ts
-
-    data = []
-    first_model = True
-    for key, value in results_dict.items():
-        aggregate_metric_dist = value[0]
-        row = [key]
-        for k, v in aggregate_metric_dist.items():
-            row.append(v)
-        if first_model:
-            data.append(["Model"] + list(aggregate_metric_dist.keys()))
-            first_model = False
-        data.append(row)
-    dataframe = pd.DataFrame(data)
-    csvname = Path(report_filename).with_suffix(".csv")
-
-    # write dataframe to csvname
-    dataframe.to_csv(csvname, index=False, header=False)
-    logger.info(f"Evaluation complete. Results saved to {csvname}")
-
-    return results_dict
-
-
-@app.command()
-def evaluate(
-    model_name: ModelType | str,
-    dataset_name: Optional[DataSetType] = None,
-    dataset_country: Optional[str] = None,
-    dataset_csv: Optional[Path] = None,
-    polygons_json: Optional[Path] = None,
-    polygons_id_field: Optional[str] = "id",
-    prediction_length: int = 6,
-    n_splits: int = 7,
-    report_filename: Optional[str] = str(get_temp_dir() / "report.pdf"),
-    ignore_environment: bool = False,
-    debug: bool = False,
-    log_file: Optional[str] = None,
-    run_directory_type: Optional[Literal["latest", "timestamp", "use_existing"]] = "timestamp",
-    model_configuration_yaml: Optional[str] = None,
-    is_chapkit_model: bool = False,
-):
-    initialize_logging(debug, log_file)
-    logger.info(f"Evaluating model {model_name}")
-    dataset = _load_dataset(dataset_country, dataset_csv, dataset_name, polygons_id_field, polygons_json)
-    model_configuration_yaml_list, model_list = _create_model_lists(model_configuration_yaml, model_name)
-    logger.info(f"Model configuration: {model_configuration_yaml_list}")
-    results_dict = {}
-    for name, configuration in zip(model_list, model_configuration_yaml_list):
-        model = _get_model(configuration, ignore_environment, is_chapkit_model, name, run_directory_type)
-        try:
-            results = evaluate_model(
-                model,
-                dataset,
-                prediction_length=prediction_length,
-                n_test_sets=n_splits,
-                report_filename=report_filename,
-            )
-        except NoPredictionsError as e:
-            logger.error(f"No predictions were made: {e}")
-            return
-        results_dict[name] = results
-
-    _save_results(report_filename, results_dict)
-
-    return results_dict
-
-
-@app.command()
-def evaluate2(
-    model_name: str,
-    dataset_csv: Path,
-    output_file: Path,
-    backtest_params: BackTestParams = BackTestParams(n_periods=3, n_splits=7, stride=1),
-    run_config: RunConfig = RunConfig(),
-    model_configuration_yaml: Optional[Path] = None,
-):
-    """
-    Evaluate a single model and export results to NetCDF format using xarray.
-
-    This command evaluates one model and outputs evaluation results in NetCDF format
-    for easy integration with scientific tools. GeoJSON polygons are automatically
-    discovered from a file with the same name as the CSV but with .geojson extension.
-
-    Args:
-        model_name: Model identifier (path or GitHub URL)
-        dataset_csv: Path to CSV file with disease data
-        output_file: Path to output NetCDF file
-        backtest_params: Backtest configuration (n_periods, n_splits, stride)
-        run_config: Model run environment configuration
-        model_configuration_yaml: Optional YAML file with model configuration
-    """
-    from chap_core.database.model_templates_and_config_tables import ConfiguredModelDB, ModelTemplateDB
-    from chap_core.log_config import initialize_logging
-
-    logger.info(f"Evaluating model {model_name} with xarray/NetCDF output")
-
-    # Initialize logging
-    initialize_logging(run_config.debug, run_config.log_file)
-
-    # Load dataset (GeoJSON auto-discovered)
-    geojson_path = _discover_geojson(dataset_csv)
-    dataset = _load_dataset_from_csv(dataset_csv, geojson_path)
-
-    # Load model configuration
-    configuration = None
-    if model_configuration_yaml is not None:
-        logger.info(f"Loading model configuration from {model_configuration_yaml}")
-        configuration = ModelConfiguration.model_validate(yaml.safe_load(open(model_configuration_yaml)))
-
-    # Load model template
-    logger.info(f"Loading model template from {model_name}")
-    template = ModelTemplate.from_directory_or_github_url(
-        model_name,
-        base_working_dir=Path("./runs/"),
-        ignore_env=run_config.ignore_environment,
-        run_dir_type=run_config.run_directory_type,
-        is_chapkit_model=run_config.is_chapkit_model,
-    )
-
-    # Configure and instantiate model
-    model = template.get_model(configuration)
-    estimator = model()
-
-    # Create configured model metadata
-    model_template_db = ModelTemplateDB(
-        id=template.model_template_config.name,
-        name=template.model_template_config.name,
-        version=template.model_template_config.version or "unknown",
-    )
-
-    configured_model_db = ConfiguredModelDB(
-        id="cli_eval",
-        model_template_id=model_template_db.id,
-        model_template=model_template_db,
-        configuration=configuration.model_dump() if configuration else {},
-    )
-
-    # Create evaluation by running backtest
-    logger.info(
-        f"Running backtest with {backtest_params.n_splits} splits, {backtest_params.n_periods} periods, stride {backtest_params.stride}"
-    )
-    evaluation = Evaluation.create(
-        configured_model=configured_model_db,
-        estimator=estimator,
-        dataset=dataset,
-        backtest_params=backtest_params,
-        backtest_name=f"{model_name}_evaluation",
-    )
-
-    # Export to file
-    logger.info(f"Exporting evaluation to {output_file}")
-    evaluation.to_file(
-        filepath=output_file,
-        model_name=model_name,
-        model_configuration=configuration.model_dump() if configuration else {},
-        model_version=template.model_template_config.version or "unknown",
-    )
-
-    logger.info(f"Evaluation complete. Results saved to {output_file}")
-
-
-def _get_model(
-    configuration: str,
-    ignore_environment: bool,
-    is_chapkit_model: bool,
-    name,
-    run_directory_type: Literal["latest", "timestamp", "use_existing"] | None,
-) -> Any:
-    template = ModelTemplate.from_directory_or_github_url(
-        name,
-        base_working_dir=Path("./runs/"),
-        ignore_env=ignore_environment,
-        run_dir_type=run_directory_type,
-        is_chapkit_model=is_chapkit_model,
-    )
-    logger.info(f"Model template loaded: {template}")
-    if configuration is not None:
-        logger.info(f"Loading model configuration from yaml file {configuration}")
-        configuration = ModelConfiguration.model_validate(
-            yaml.safe_load(open(configuration))
-        )  # template.get_model_configuration_from_yaml(Path(configuration))
-        logger.info(f"Loaded model configuration from yaml file: {configuration}")
-
-    model = template.get_model(configuration)
-    model = model()
-    return model
-
-
-def _save_results(report_filename: str | None, results_dict: dict[Any, Any]):
-    # need to iterate through the dict, like key and value or something and then extract the relevant metrics
-    # to a pandas dataframe, and save it as a csv file.
-    # it seems like results contain two dictionairies, one for aggregate metrics and one with seperate ones for each ts
-
-    data = []
-    full_data = {}
-    first_model = True
-    for key, value in results_dict.items():
-        aggregate_metric_dist = value[0]
-        row = [key]
-        for k, v in aggregate_metric_dist.items():
-            row.append(v)
-        if first_model:
-            data.append(["Model"] + list(aggregate_metric_dist.keys()))
-            first_model = False
-        data.append(row)
-        # make a dataframe with column names forst and then the full value[1]
-        full_data[key] = pd.DataFrame(value[1])
-
-    dataframe = pd.DataFrame(data)
-    csvname = Path(report_filename).with_suffix(".csv")
-    for i, (model_name, results) in enumerate(full_data.items()):
-        csvname_full = Path(report_filename).with_suffix(f".{i}.csv")
-        results.to_csv(csvname_full, index=False)
-        logger.info(f"Wrote detailed results for {model_name} to {csvname_full}")
-
-    # write dataframe to csvname
-    dataframe.to_csv(csvname, index=False, header=False)
-    logger.info(f"Evaluation complete. Results saved to {csvname}")
-
-
-def _create_model_lists(model_configuration_yaml: str | None, model_name) -> tuple[list[str], Any]:
-    if "," in model_name:
-        # model_name is not only one model, but contains a list of models
-        model_list = model_name.split(",")
-        model_configuration_yaml_list = [None for _ in model_list]
-        if model_configuration_yaml is not None:
-            model_configuration_yaml_list = model_configuration_yaml.split(",")
-            assert len(model_list) == len(model_configuration_yaml_list), (
-                "Number of model configurations does not match number of models"
-            )
-    else:
-        model_list = [model_name]
-        model_configuration_yaml_list = [model_configuration_yaml]
-    return model_configuration_yaml_list, model_list
-
-
-def _discover_geojson(csv_path: Path) -> Optional[Path]:
-    """
-    Discover GeoJSON file alongside CSV file.
-
-    Looks for a file with the same name as the CSV but with .geojson extension.
-
-    Args:
-        csv_path: Path to CSV file
-
-    Returns:
-        Path to GeoJSON file if it exists, None otherwise
-    """
-    geojson_path = Path(str(csv_path).replace(".csv", ".geojson"))
-    if geojson_path.exists():
-        return geojson_path
-    return None
-
-
-def _load_dataset_from_csv(csv_path: Path, geojson_path: Optional[Path] = None) -> DataSet:
-    """
-    Load dataset from CSV file with optional GeoJSON polygons.
-
-    Args:
-        csv_path: Path to CSV file with disease data
-        geojson_path: Optional path to GeoJSON file with polygon boundaries
-
-    Returns:
-        DataSet loaded from CSV with polygons if provided
-    """
-    logging.info(f"Loading dataset from {csv_path}")
-    dataset = DataSet.from_csv(csv_path, FullData)
-
-    if geojson_path is not None:
-        logging.info(f"Loading polygons from {geojson_path}")
-        polygons = Polygons.from_file(geojson_path, id_property="id")
-        polygons.filter_locations(dataset.locations())
-        dataset.set_polygons(polygons.data)
-
-    return dataset
-
-
-def _load_dataset(
-    dataset_country: str | None,
-    dataset_csv: Path | None,
-    dataset_name: Any | None,
-    polygons_id_field: str | None,
-    polygons_json: Path | None,
-) -> DataSet:
-    if dataset_name is None:
-        assert dataset_csv is not None, "Must specify a dataset name or a dataset csv file"
-        logging.info(f"Loading dataset from {dataset_csv}")
-        dataset = DataSet.from_csv(dataset_csv, FullData)
-        if polygons_json is not None:
-            logging.info(f"Loading polygons from {polygons_json}")
-            polygons = Polygons.from_file(polygons_json, id_property=polygons_id_field)
-            polygons.filter_locations(dataset.locations())
-            dataset.set_polygons(polygons.data)
-    else:
-        logger.info(f"Evaluating model on dataset {dataset_name}")
-        dataset = datasets[dataset_name]
-        dataset = dataset.load()
-
-        if isinstance(dataset, MultiCountryDataSet):
-            assert dataset_country is not None, "Must specify a country for multi country datasets"
-            assert dataset_country in dataset.countries, (
-                f"Country {dataset_country} not found in dataset. Countries: {dataset.countries}"
-            )
-            dataset = dataset[dataset_country]
-    return dataset
-
-
-@app.command()
-def sanity_check_model(
-    model_url: str, use_local_environement: bool = False, dataset_path=None, model_config_path: str = None
-):
-    """
-    Check that a model can be loaded, trained and used to make predictions
-    """
-    if dataset_path is None:
-        dataset = datasets["hydromet_5_filtered"].load()
-    else:
-        dataset = DataSet.from_csv(dataset_path, FullData)
-    train, tests = train_test_generator(dataset, 3, n_test_sets=2)
-    context, future, truth = next(tests)
-    logger.info("Dataset: ")
-    logger.info(dataset.to_pandas())
-
-    if model_config_path is not None:
-        model_config = ModelConfiguration.model_validate(yaml.safe_load(open(model_config_path)))
-    else:
-        model_config = None
-    try:
-        model_template = get_model_template_from_directory_or_github_url(model_url, ignore_env=use_local_environement)
-        model = model_template.get_model(
-            model_config
-        )  # get_model_from_directory_or_github_url(model_url, ignore_env=use_local_environement)
-        estimator = model()
-    except Exception as e:
-        logger.error(f"Error while creating model: {e}")
-        raise e
-    try:
-        predictor = estimator.train(train)
-    except Exception as e:
-        logger.error(f"Error while training model: {e}")
-        raise e
-    try:
-        predictions = predictor.predict(context, future)
-    except Exception as e:
-        logger.error(f"Error while forecasting: {e}")
-        raise e
-    for location, prediction in predictions.items():
-        assert not np.isnan(prediction.samples).any(), (
-            f"NaNs in predictions for location {location}, {prediction.samples}"
-        )
-    context, future, truth = next(tests)
-    try:
-        predictions = predictor.predict(context, future)
-    except Exception as e:
-        logger.error(f"Error while forecasting from a future time point: {e}")
-        raise e
-    for location, prediction in predictions.items():
-        assert not np.isnan(prediction.samples).any(), (
-            f"NaNs in futuresplit predictions for location {location}, {prediction.samples}"
-        )
-
-
-@app.command()
-def forecast(
-    model_name: str,
-    dataset_name: DataSetType,
-    n_months: int,
-    model_path: Optional[str] = None,
-    out_path: Optional[str] = Path("./"),
-):
-    """
-    Forecast n_months ahead using the given model and dataset
-
-    Parameters:
-        model_name: Name of the model to use, set to external to use an external model and specify the external model with model_path
-        dataset_name: Name of the dataset to use, e.g. hydromet_5_filtered
-        n_months: int: Number of months to forecast ahead
-        model_path: Optional[str]: Path to the model if model_name is external. Can ge a github repo url starting with https://github.com and ending with .git or a path to a local directory.
-        out_path: Optional[str]: Path to save the output file, default is the current directory
-    """
-
-    out_file = Path(out_path) / f"{model_name}_{dataset_name}_forecast_results_{n_months}.html"
-    f = open(out_file, "w")
-    figs = api.forecast(model_name, dataset_name, n_months, model_path)
-    for fig in figs:
-        f.write(fig.to_html())
-    f.close()
-
-
-@app.command()
-def multi_forecast(
-    model_name: str,
-    dataset_name: DataSetType,
-    n_months: int,
-    pre_train_months: int,
-    out_path: Path = Path(""),
-):
-    model = get_model_from_directory_or_github_url(model_name)
-    model_name = model.name
-
-    model = model()
-    filename = out_path / f"{model_name}_{dataset_name}_multi_forecast_results_{n_months}.html"
-    logger.info(f"Saving to {filename}")
-    f = open(filename, "w")
-    dataset = datasets[dataset_name].load()
-    predictions_list = list(
-        do_multi_forecast(
-            model,
-            dataset,
-            n_months * delta_month,
-            pre_train_delta=pre_train_months * delta_month,
-        )
-    )
-
-    for location, true_data in dataset.items():
-        local_predictions = [pred.get_location(location).data() for pred in predictions_list]
-        fig = plot_forecast_from_summaries(local_predictions, true_data.data())
-        f.write(fig.to_html())
-    f.close()
-
-
-@app.command()
-def serve(seedfile: Optional[str] = None, debug: bool = False, auto_reload: bool = False):
-    """
-    Start CHAP as a backend server
-    """
-    from .rest_api_src.v1.rest_api import main_backend
-
-    logger.info("Running chap serve")
-
-    if seedfile is not None:
-        data = json.load(open(seedfile))
-    else:
-        data = None
-
-    main_backend(data, auto_reload=auto_reload)
-
-
-@app.command()
-def write_open_api_spec(out_path: str):
-    """
-    Write the OpenAPI spec to a file
-    """
-    from chap_core.rest_api_src.v1.rest_api import get_openapi_schema
-
-    schema = get_openapi_schema()
-    with open(out_path, "w") as f:
-        json.dump(schema, f, indent=4)
-
-
-def base_args(func, *args, **kwargs):
-    """
-    Decorator that adds some base arguments to a command
-    """
-    base_args = [("debug", bool, False), ("log_file", Optional[str], None)]
-
-    def new_func(*args, **kwargs):
-        for arg_name, arg_type, default in base_args:
-            if arg_name not in kwargs:
-                kwargs[arg_name] = default
-        return func(*args, **kwargs)
-
-
-@app.command()
-def test(**base_kwargs):
-    """
-    Simple test-command to check that the chap command works
-    """
-    initialize_logging()
-
-    logger.debug("Debug message")
-    logger.info("Info message")
-
-
-@dataclasses.dataclass
-class AreaPolygons: ...
-
-
-@app.command()
-def plot_dataset(data_filename: Path, plot_name: str = "standardized_feature_plot"):
-    dataset_plot_registry = {
-        "standardized_feature_plot": StandardizedFeaturePlot,
-        "season_plot": SeasonCorrelationBarPlot,
-    }
-    plot_cls = dataset_plot_registry.get(plot_name, StandardizedFeaturePlot)
-    df = pd.read_csv(data_filename)
-    plotter = plot_cls(df)
-    fig = plotter.plot()
-    fig.show()
-
-
-def main_function():
-    """
-    This function should just be type hinted with common types,
-    and it will run as a command line function
-    Simple function<
-
-    >>> main()
-
-    """
-    return
+# Register commands from each module
+evaluate.register_commands(app)
+forecast.register_commands(app)
+utils.register_commands(app)
 
 
 def main():
