+++ chap_core/cli_endpoints/__init__.py
@@ -1,0 +1,8 @@
+"""CLI endpoints for CHAP Core.
+
+This package contains the CLI command implementations, organized by functionality.
+"""
+
+from chap_core.cli_endpoints import evaluate, forecast, utils
+
+__all__ = ["evaluate", "forecast", "utils"]
